+++ packages/gamut-styles/src/theme/utils/shouldForwardProp.ts
@@ -1,0 +1,13 @@
+import isPropValid from '@emotion/is-prop-valid';
+
+import { properties } from '../props';
+
+const allProps = Object.keys(properties).reduce<string[]>(
+  (carry, prop: keyof typeof properties) => {
+    return [...carry, ...properties[prop].propNames];
+  },
+  []
+);
+
+export const shouldForwardProp = (prop: string) =>
+  isPropValid(prop) && !allProps.includes(prop);